+++ Laborki/mobile/app/src/main/AndroidManifest.xml
@@ -9,11 +9,10 @@
         android:name="pl.lipov.laborki.App"
         android:allowBackup="true"
         android:fullBackupContent="false"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@drawable/app_logo"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme"
+        android:theme="@style/SplashScreen"
         tools:ignore="GoogleAppIndexingWarning">
 
         <activity
