+++ Laborki/mobile/app/src/main/java/pl/lipov/laborki/common/utils/GestureDetectorUtils.kt
@@ -1,6 +1,7 @@
 package pl.lipov.laborki.common.utils
 
 import android.view.GestureDetector
+import android.view.MotionEvent
 import androidx.lifecycle.MutableLiveData
 import pl.lipov.laborki.data.model.Event
 
@@ -8,4 +9,14 @@ class GestureDetectorUtils : GestureDetector.SimpleOnGestureListener() {
 
     val onEvent = MutableLiveData<Event>()
     private var gestureDetector: GestureDetector? = null
+
+    override fun onLongPress(e: MotionEvent?) {
+        onEvent.postValue(Event.LONG_CLICK)
+        super.onLongPress(e)
+    }
+
+    override fun onDoubleTap(e: MotionEvent?): Boolean {
+        onEvent.postValue(Event.DOUBLE_TAP)
+        return super.onDoubleTap(e)
+    }
 }
