+++ Laborki/mobile/app/src/main/java/pl/lipov/laborki/presentation/login/AuthCallback.kt
@@ -1,0 +1,5 @@
+package pl.lipov.laborki.presentation.login
+
+interface AuthCallback {
+    fun onLoginSuccess()
+}
