+++ Laborki/mobile/app/src/main/java/pl/lipov/laborki/presentation/login/LoginCallback.kt
@@ -1,5 +1,5 @@
 package pl.lipov.laborki.presentation.login
 
 interface LoginCallback {
-    fun onLoginSuccess()
+//    fun onLoginSuccess()
 }
