+++ Laborki/mobile/app/src/main/res/drawable-v24/app_logo.png
