+++ Laborki/mobile/app/src/main/res/drawable/main_background_gradient.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="90"
+        android:endColor="@color/end_color"
+        android:centerColor="@color/center_color"
+        android:startColor="@android:color/white"
+        android:type="linear" />
+</shape>
