+++ Laborki/mobile/app/src/main/res/drawable/pw.jpg
