+++ Laborki/mobile/app/src/main/res/layout/activity_main.xml
@@ -1,87 +1,55 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    >
+    android:background="@android:color/white"
+    android:id="@+id/activity_main"
+    tools:context=".presentation.MainActivity">
 
-    <View
-        android:id="@+id/ic_bigstar"
-        android:layout_width="@dimen/big_icon_size"
-        android:layout_height="@dimen/big_icon_size"
-        android:background="@drawable/ic_bigstar"
-        android:contentDescription="@null"
-        android:visibility="visible"
-        android:layout_marginBottom="24dp"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="110dp"/>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:gravity="center"
+        android:background="@android:color/white"
+        android:gravity="center_horizontal"
         android:orientation="vertical">
 
-        <TextView
+        <ImageView
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:text="@string/gestures_text"
-            android:textSize="24sp"
-            android:layout_marginBottom="20dp"/>
+            android:layout_height="220dp"
+            android:contentDescription="@null"
+            android:scaleType="centerCrop"
+            app:srcCompat="@drawable/pw"/>
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:orientation="horizontal">
 
-            <View
-                android:id="@+id/ic_star1"
-                android:layout_width="@dimen/lock_icon_size"
-                android:layout_height="@dimen/lock_icon_size"
-                android:background="@drawable/ic_star"
-                android:contentDescription="@null" />
-
-            <View
-                android:id="@+id/ic_star2"
-                android:layout_width="@dimen/lock_icon_size"
-                android:layout_height="@dimen/lock_icon_size"
-                android:background="@drawable/ic_star"
-                android:contentDescription="@null" />
-
-            <View
-                android:id="@+id/ic_star3"
-                android:layout_width="@dimen/lock_icon_size"
-                android:layout_height="@dimen/lock_icon_size"
-                android:background="@drawable/ic_star"
-                android:contentDescription="@null" />
-
-            <View
-                android:id="@+id/ic_star4"
-                android:layout_width="@dimen/lock_icon_size"
-                android:layout_height="@dimen/lock_icon_size"
-                android:background="@drawable/ic_star"
-                android:contentDescription="@null" />
-
-        </LinearLayout>
-
-        <Button
-            android:id="@+id/login_button"
+        <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/login_button_text"
-            android:layout_marginTop="25dp"/>
+            android:contentDescription="@null"
+            android:scaleType="centerCrop"
+            app:srcCompat="@drawable/app_logo"/>
+
+        <FrameLayout
+            android:id="@+id/fragment_container"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_gravity="center"/>
 
     </LinearLayout>
 
-</RelativeLayout>
+    <ImageView
+        android:layout_width="match_parent"
+        android:layout_height="220dp"
+        android:alpha="0.99"
+        android:contentDescription="@null"
+        android:scaleType="centerCrop"
+        app:srcCompat="@drawable/main_background_gradient"/>
 
-<!--    <?xml version="1.0" encoding="utf-8"?>-->
-<!--<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"-->
-<!--android:id="@+id/fragment_container"-->
-<!--android:layout_width="match_parent"-->
-<!--android:layout_height="match_parent"-->
-<!--android:gravity="center"-->
-<!--android:orientation="vertical">-->
-<!--</FrameLayout>-->
+    <FrameLayout
+        android:id="@+id/fragment_authentication"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
+
+</RelativeLayout>
