+++ Laborki/mobile/app/src/main/res/layout/fragment_login.xml
@@ -1,16 +1,80 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout>
-
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:gravity="center"
-        android:orientation="vertical">
+        android:id="@+id/fragment_login">
 
-        <Button
-            android:id="@+id/login_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Login" />
-    </LinearLayout>
+        <View
+            android:id="@+id/ic_bigstar"
+            android:layout_width="24dp"
+            android:layout_height="24dp"
+            android:background="@drawable/ic_bigstar"
+            android:contentDescription="@null"
+            android:visibility="visible"
+            android:layout_marginBottom="24dp"
+            android:layout_centerHorizontal="true"
+            android:layout_marginTop="50dp" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:gravity="center"
+            android:orientation="vertical">
+
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:gravity="center"
+                android:text="@string/gestures_text"
+                android:textSize="24sp"
+                android:layout_marginBottom="20dp"/>
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:gravity="center"
+                android:orientation="horizontal">
+
+                <View
+                    android:id="@+id/ic_star1"
+                    android:layout_width="@dimen/lock_icon_size"
+                    android:layout_height="@dimen/lock_icon_size"
+                    android:background="@drawable/ic_star"
+                    android:contentDescription="@null" />
+
+                <View
+                    android:id="@+id/ic_star2"
+                    android:layout_width="@dimen/lock_icon_size"
+                    android:layout_height="@dimen/lock_icon_size"
+                    android:background="@drawable/ic_star"
+                    android:contentDescription="@null" />
+
+                <View
+                    android:id="@+id/ic_star3"
+                    android:layout_width="@dimen/lock_icon_size"
+                    android:layout_height="@dimen/lock_icon_size"
+                    android:background="@drawable/ic_star"
+                    android:contentDescription="@null" />
+
+                <View
+                    android:id="@+id/ic_star4"
+                    android:layout_width="@dimen/lock_icon_size"
+                    android:layout_height="@dimen/lock_icon_size"
+                    android:background="@drawable/ic_star"
+                    android:contentDescription="@null" />
+
+            </LinearLayout>
+
+            <Button
+                android:id="@+id/login_button"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/login_button_text"
+                android:layout_marginTop="25dp"/>
+
+        </LinearLayout>
+
+    </RelativeLayout>
 </layout>
