+++ Laborki/mobile/app/src/main/res/values/colors.xml
@@ -2,7 +2,10 @@
 <resources>
     <color name="colorPrimary">#0099ff</color>
     <color name="colorPrimaryDark">#004e82</color>
+    <color name="primaryDarkAlt">#2b2926</color>
     <color name="colorAccent">#D81B60</color>
+    <color name="end_color">#414241</color>
+    <color name="center_color">#99000000</color>
     <color name="army">#4b5320</color>
     <color name="army_dark">#3e451b</color>
     <color name="grey">#bfbfbf</color>
