+++ Laborki/mobile/app/src/main/res/values/strings.xml
@@ -2,4 +2,6 @@
     <string name="app_name">Laborki</string>
     <string name="login_button_text">Login</string>
     <string name="gestures_text">Use gestures to unlock</string>
+    <string name="login_text">Type in your login</string>
+    <string name="login_hint">Your login...</string>
 </resources>
